Fix build to properly separate desktop and moonlight profile bits.
I was playing things too fast and loose before and somehow they were magically working. But the
latest version of MonoTouch is not so lenient. So IKVM now has to be built explicitly in five
different phases:

Phase 1, 3, and 5 are built with "nant" (and hence the desktop profile). Phase 2 and 4 are built
with "nant -t:moonlight-2.0" (the silverlight/moonlight/monotouch profile). Tragically it is now
also necessary to modify the NAnt.exe.config file to get the -t:moonlight-2.0 parts to build.